Move core_preferences to proto package
