+++ app/src/main/java/com/sduduzog/slimlauncher/ui/dialogs/ChooseClockTypeDialog.kt
@@ -4,7 +4,7 @@ import android.app.AlertDialog
 import android.app.Dialog
 import android.os.Bundle
 import androidx.fragment.app.DialogFragment
-import com.jkuester.unlauncher.datastore.ClockType
+import com.jkuester.unlauncher.datastore.proto.ClockType
 import com.sduduzog.slimlauncher.R
 import com.sduduzog.slimlauncher.datasource.coreprefs.CorePreferencesRepository
 import dagger.hilt.android.AndroidEntryPoint
