+++ app/src/main/proto/core_preferences.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option java_package = "com.jkuester.unlauncher.datastore";
+option java_package = "com.jkuester.unlauncher.datastore.proto";
 option java_multiple_files = true;
 
 message CorePreferences {
