+++ app/src/test/java/com/jkuester/unlauncher/datasource/DataSourceModuleTest.kt
@@ -2,7 +2,7 @@ package com.jkuester.unlauncher.datasource
 
 import android.content.Context
 import androidx.datastore.core.DataStore
-import com.jkuester.unlauncher.datastore.CorePreferences
+import com.jkuester.unlauncher.datastore.proto.CorePreferences
 import com.jkuester.unlauncher.datastore.proto.QuickButtonPreferences
 import com.jkuester.unlauncher.datastore.proto.UnlauncherApps
 import io.mockk.every
